+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.0.1
+## 0.2.2
+* Initial release
 
-* TODO: Describe initial release.
+## 0.2.2+2
+* User can enter footer title without having to enter footer icon
+
+## 0.2.2+5
+* Update default theme data